Only draw the Display's selection extent frame if it exists.
The JS interpreter was raising an exception every time
through the draw loop because it was calling the 'as' method
with this.extent as the argument, but the value was undefined.

I thought that this fix would result in less CPU usage
while Safari/Phosphor was idle, but alas, no joy.