Changed Phosphor's Names object to just Name.
This may be personal taste, but I interpreted Names.of(this.content)
to mean that it would return a list of names, but in fact it returns
a single string. This is especially confusing since Object.prototype.names()
actually does return a list of names.

One additional thing that I considered here was creating some syntax
for conditionally binding a value to a key iff the value is defined.

For example, phosphor.js line 138 could read:
	this.conditionallyBind({content: Name.of(this.content)});
where
	Object.prototype.conditionallyBind = function(o) {
		var $self = this;
		o.each(function (v,k) { if (v != undefined) $self[k] = v });
		return this;
	}
It's basically copy() unless v is undefined. The only reason I didn't do
it is that there were no other parts of the program that jumped out at
me shouting, "Please, conditionallyBind me."