Added facets(): introspect an object's properties by prototype group.
The motivation behind the facets function is to group an object's properties
by the object that it inherited from (although this is not stricly true due
to the paucity of prototpe metadata at runtime). Such a grouped list can then
be pretty-printed to give the user a better understanding of an object's capabilities.

I'd like to make this grouped property presentation available via the Phosphor GUI,
but for now it lives in a REPL world:
REPL> pprint(facets(Mouse), 0)

Originally I implemented this in terms of the 'implements()' method,
but I pulled the body of that method into here in order to collapse the loop.

Note: there are some problems with the current is() implementation. For example:
1. Circle.is(Name)      => why: both have an 'of' property, although different semantic meaning.
2. Circle.is(Territory) => why: Territory is a superset of Circle.
3. Circle.is(Event)     => why: Event only defines 'init', and Circle also has 'init'.

In order for the facets function to become useful, these issues in the 'is()'
implementation would need to be worked out.